shop/requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-elasticsearch-dsl==7.2.2
django-filter==21.1
django-timezone-field==4.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
kombu==5.2.2
Pillow==8.4.0
prompt-toolkit==3.0.24
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
python-dateutil==2.8.2
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
psycopg2==2.9.3
